stack trace requests now operational
