+++ test/cppworkspace/test/testspawn.sh
@@ -1,0 +1,1 @@
+gdb build/testapp -iex "source ../../../modules/python/setup.py" -iex "source ../../../modules/python/stdlib.py"
